+++ lib/datamappify/data/criteria/active_record/limit.rb
@@ -2,7 +2,13 @@ module Datamappify
   module Data
     module Criteria
       module ActiveRecord
-        class Limit < Relational::Limit
+        class Limit < Common
+          def records(scope = nil)
+            limit, offset = criteria
+            offset ||= 0
+
+            (scope || source_class).limit(limit).offset(offset)
+          end
         end
       end
     end