+++ lib/datamappify/data/criteria/sequel/limit.rb
@@ -2,7 +2,10 @@ module Datamappify
   module Data
     module Criteria
       module Sequel
-        class Limit < Relational::Limit
+        class Limit < Common
+          def records(scope = nil)
+            (scope || source_class).limit(*criteria)
+          end
         end
       end
     end