+++ spec/repository/finders/criteria_spec.rb
@@ -3,6 +3,7 @@ require 'spec_helper'
 shared_examples_for "finders examples" do |data_provider|
   [
     { :where => { :last_name => 'Superman' }, :limit => 2, :order => { :last_name => :asc, :first_name => :desc, :id => :desc } },
+    { :where => { :last_name => 'Superman' }, :limit => [2, 0], :order => { :last_name => :asc, :first_name => :desc, :id => :desc } },
     { :order => { :last_name => :asc, :first_name => :desc, :id => :desc }, :limit => 2, :where => { :last_name => 'Superman' } },
     { :where => {}, :match => { :last_name => 'Super%' }, :limit => 2, :order => { :last_name => :asc, :first_name => :desc, :id => :desc } }
   ].each_with_index do |criteria, index|